Security: Prevent re-creation of a Data Source
**Before**: When a Data Source was created and the corresponding
S3 Bucket was already present we accepted the exception and allow its
re-creation in the Control Panel. Admin could then access the data
and grand access to other users. This is not ideal because the S3 Bucket
could contain data.

**Now**: When a user creates a Data Source and there is already an S3
bucket with the same name its creation in the Control Panel is aborted
and an error is shown to the user. This prevent this user access to
potential data already present on this S3 Bucket.

**TODO**: Currently not handling exception when creating S3 buckets
for an app

Ticket: https://trello.com/c/vS8iQzoP