Add "worker_health" command to check worker health
This command will be used by Kubernetes to check on the Django Channels
worker using a "command" probe:

https://kubernetes.io/docs/tasks/configure-pod-container/configure-liveness-readiness-startup-probes/#define-a-liveness-command

How does it work?
=================

The idea behind this is that the command will submit a "worker_health"
background job (asyncronous) and then check the existance/freshness of
the `/tmp/worker_health.txt` file.

As:
1. kubernetes will periodically execute this command and
2. the command will periodically submit this "heathbeat" `worker_health`
   background job

this means that the file should be almost always relatively fresh (
freshness may need tweaking using `--seconds` argument).

There are some subtleties to bear in mind:
1. there will be several instances of the workers (currently 3)
2. a background task submitted on a pod may be executed on a different pod
2a. This means that on average you can expect a given worker to process
    1/n of the tasks (and this needs to be backed into the `--seconds` and
    k8s probe configuration)
3. the worker probe command run on the same container where the worker
   health task will run, hence using a local, temporary file is
   satisfactory as a way for these to "communicate" (and it's very simple
   and understandable)
4. 2a and 3 above means that even if a health task is processed correctly
   every `--seconds` seconds the actual file touched may not be the same
   (again, this is OK but will need to be baked into the k8s probe
   configuration)
5. From local testing the worker health task is processed very quickly
   but because of the asyncronous nature of these background tasks and
   the existance of several workers on independent containers processing
   them it means that probe will need to be tweaked in a certain way
   so that few failures are acceptable

What the above mean is that the kubernetes liveliness probe will need to
be configured so that few failures are acceptable and it would not be
too aggressive into killing a worker container because of false positives.
(this is in general a best practice when it comes to liveliness probes,
configuring them is part science and part art).

Part of ticket: https://trello.com/c/eUMRw1gK